+++ extension-style-lint-rules/src/test/java/com/mapbox/maps/lint/style/Stubs.kt
@@ -1,4 +1,4 @@
-package com.mapbox.maps.lint.expressions
+package com.mapbox.maps.lint.style
 
 import com.android.tools.lint.checks.infrastructure.LintDetectorTest.kotlin
 import com.android.tools.lint.checks.infrastructure.TestFile
@@ -228,5 +228,186 @@ object Stubs {
       }
 
     """.trimIndent()
+  )
+  val styleDslClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style
+      
+      import com.mapbox.maps.LayerPosition
+      import com.mapbox.maps.extension.style.layers.*
+      import com.mapbox.maps.extension.style.light.generated.Light
+      import com.mapbox.maps.extension.style.projection.generated.Projection
+      import com.mapbox.maps.extension.style.sources.*
+      import com.mapbox.maps.extension.style.terrain.generated.Terrain
+
+      class StyleExtensionImpl {
+        class Builder(styleUri: String) {
+          @JvmName("addLayer")
+          operator fun Layer.unaryPlus() {
+            // stub
+          }
+          @JvmName("addLayerAtPosition")
+          operator fun Pair<Layer, LayerPosition>.unaryPlus() {
+            // stub
+          }
+          @JvmName("addSource")
+          operator fun Source.unaryPlus() {
+            // stub
+          }
+          @JvmName("setLight")
+          operator fun Light.unaryPlus() {
+            // stub
+          }
+          @JvmName("setTerrain")
+          operator fun Terrain.unaryPlus() {
+            // stub
+          }
+          @JvmName("setAtmosphere")
+          operator fun Atmosphere.unaryPlus() {
+            // stub
+          }
+          @JvmName("setProjection")
+          operator fun Projection.unaryPlus() {
+            // stub
+          }
+          @JvmName("addImage")
+          operator fun ImageExtensionImpl.unaryPlus() {
+            // stub
+          }
+          @JvmName("addImage9Patch")
+          operator fun ImageNinePatchExtensionImpl.unaryPlus() {
+            // stub
+          }
+          @JvmOverloads
+          fun layerAtPosition(
+            layer: Layer,
+            above: String? = null,
+            below: String? = null,
+            at: Int? = null
+          ): Pair<Layer, LayerPosition> {
+            // stub
+            TODO()
+          }
+          fun build(): StyleExtensionImpl {
+            return StyleExtensionImpl()
+          }
+        }
+      }
+      fun style(styleUri: String = "", block: StyleExtensionImpl.Builder.() -> Unit): StyleExtensionImpl =
+        StyleExtensionImpl.Builder(styleUri).apply(block).build()
+    """.trimIndent()
+  )
+  val layerClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.layers
+      abstract class Layer {
+        // stub
+        abstract fun stub()
+      }
+    """.trimIndent()
+  )
+  val layerPositionClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps
+      class LayerPosition {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val symbolLayerClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.layers.generated
+      
+      import com.mapbox.maps.extension.style.layers.Layer
+
+      class SymbolLayer(val layerId: String, val sourceId: String) : Layer() {
+        // stub
+        override fun stub() {
+          // stub
+        }
+      }
+    """.trimIndent()
+  )
+  val sourceClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.sources.generated
+      abstract class Source {
+        // stub
+        abstract fun stub()
+      }
+    """.trimIndent()
+  )
+  val geoJsonSourceClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.sources.generated
+      class GeoJsonSource(val sourceId: String) : Source() {
+        // stub
+        override fun stub() {
+          // stub
+        }
+      }
+    """.trimIndent()
+  )
+  val lightClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.light.generated
+      class Light {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val atmosphereClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.atmosphere.generated
+      class Atmosphere {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val terrainClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.terrain.generated
+      class Terrain {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val projectionClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.projection.generated
+      class Projection(val name: String) {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val imageClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.image
+      class ImageExtensionImpl {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val imageNinePatchClassFile: TestFile = kotlin(
+    """
+      package com.mapbox.maps.extension.style.image
+      class ImageNinePatchExtensionImpl {
+        // stub
+      }
+    """.trimIndent()
+  )
+  val styleDslTestFiles = arrayOf(
+    styleDslClassFile,
+    layerClassFile,
+    layerPositionClassFile,
+    symbolLayerClassFile,
+    sourceClassFile,
+    geoJsonSourceClassFile,
+    lightClassFile,
+    atmosphereClassFile,
+    terrainClassFile,
+    projectionClassFile,
+    imageClassFile,
+    imageNinePatchClassFile
   )
 }